Refactored code
I have also taken the opportunity to change around some documentation and error messages, and made everything work together a little bit better.

+++ src/application.cpp
@@ -71,7 +71,7 @@ namespace Engine {
             camera.setPerspectiveProjection(glm::radians(50.0f), aspectRatio, 0.1f, 100.0f);
 
             if (auto commandBuffer = renderer.beginFrame()) {
-                int frameIndex = renderer.getCurrentFrameIndex();
+                size_t frameIndex = renderer.getCurrentFrameIndex();
                 FrameInfo frameInfo{frameIndex,
                                     deltaTime,
                                     commandBuffer,
